% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service.R
\name{createService}
\alias{createService}
\title{Create a pushjet.io service.}
\usage{
createService(name, icon = "https://www.r-project.org/favicon-32x32.png")
}
\arguments{
\item{name}{name of the service}

\item{icon}{icon of the service (default is the R-Project icon)}
}
\value{
list with service information
}
\description{
Create a pushjet.io service.
}

