% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service.R
\name{deleteService}
\alias{deleteService}
\title{Remove pushjet.io service.}
\usage{
deleteService(secret)
}
\arguments{
\item{secret}{secret key of service}
}
\value{
TRUE/FALSE
}
\description{
Remove pushjet.io service.
}

