% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.R
\name{pjmsg}
\alias{pjmsg}
\title{Send a message via pushjet.io}
\usage{
pjmsg(msg, title = NULL, level = 3L,
  secret = getOption("pushjetr.secret"))
}
\arguments{
\item{msg}{message}

\item{title}{message title}

\item{level}{the importance level from 1(low) to 5(high)}

\item{secret}{channel secret}
}
\value{
TRUE/FALSE
}
\description{
This function sends a message using the pushjet.io API.
}

