% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service.R
\name{serviceInfo}
\alias{serviceInfo}
\title{Get information about a pushjet.io service.}
\usage{
serviceInfo(secret = getOption("pushjetr.secret"))
}
\arguments{
\item{secret}{secret key of service}
}
\value{
list with service information
}
\description{
Get information about a pushjet.io service.
}

